@@ -1,11 +1,11 @@
 [package]
-name = "tauri-plugin-deep-link"
-version = "0.1.2"
-authors = ["FabianLars <fabianlars@fabianlars.de>"]
-description = "A Tauri plugin for deep linking support"
-repository = "https://github.com/FabianLars/tauri-plugin-deep-link"
+name = "cargo-deep-link"
+version = "0.1.0"
+authors = ["FabianLars <fabianlars@fabianlars.de>", "jf908"]
+description = "A library for deep linking support"
+repository = "https://github.com/jf908/cargo-deep-link"
 edition = "2021"
-rust-version = "1.64"
+rust-version = "1.81"
 license = "MIT OR Apache-2.0"
 readme = "README.md"
 include = ["src/**", "Cargo.toml", "LICENSE_*"]
@@ -13,8 +13,6 @@ include = ["src/**", "Cargo.toml", "LICENSE_*"]
 [dependencies]
 dirs = "5"
 log = "0.4"
-once_cell = "1"
-tauri-utils = { version = "1" }
 
 [target.'cfg(windows)'.dependencies]
 interprocess = { version = "1.2", default-features = false }
@@ -27,3 +25,7 @@ winreg = "0.55.0"
 
 [target.'cfg(target_os = "macos")'.dependencies]
 objc2 = "0.4.1"
+
+[dev-dependencies]
+eframe = "0.30"
+env_logger = "0.11"
