@@ -2,7 +2,7 @@ use std::{
     fs::remove_file,
     io::{ErrorKind, Read, Result, Write},
     os::unix::net::{UnixListener, UnixStream},
-    sync::Mutex,
+    sync::{Mutex, OnceLock},
 };
 
 use objc2::{
@@ -12,14 +12,13 @@ use objc2::{
     runtime::{AnyObject, NSObject},
     sel, ClassType,
 };
-use once_cell::sync::OnceCell;
 
 use crate::ID;
 
-type THandler = OnceCell<Mutex<Box<dyn FnMut(String) + Send + 'static>>>;
+type THandler = OnceLock<Mutex<Box<dyn FnMut(String) + Send + 'static>>>;
 
 // If the Mutex turns out to be a problem, or FnMut turns out to be useless, we can remove the Mutex and turn FnMut into Fn
-static HANDLER: THandler = OnceCell::new();
+static HANDLER: THandler = OnceLock::new();
 
 pub fn register<F: FnMut(String) + Send + 'static>(_scheme: &str, handler: F) -> Result<()> {
     listen(handler)?;
